--- a/redis_clone/parser/base.py
+++ b/redis_clone/parser/base.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+from typing import Tuple, List, AnyStr
+
+class BaseParser(ABC):
+    @abstractmethod
+    def parse(self, data, *args, **kwargs) -> Tuple[AnyStr, List[AnyStr]]:
+        pass