--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -1,5 +1,6 @@
 async-timeout==4.0.3
 exceptiongroup==1.1.3
+hiredis==2.3.2
 iniconfig==2.0.0
 packaging==23.2
 pluggy==1.3.0