Changes are not visible.

```bash
HIredis parser
❯ redis-benchmark -p 9999 -t SET,GET -q --insecure
WARNING: Could not fetch server CONFIG
SET: 11548.68 requests per second, p50=3.567 msec                   
GET: 12729.12 requests per second, p50=3.463 msec                   


My Custom Parser
❯ redis-benchmark -p 9999 -t SET,GET -q --insecure
WARNING: Could not fetch server CONFIG
SET: 10593.22 requests per second, p50=3.671 msec                   
GET: 12351.78 requests per second, p50=3.535 msec        
```